dev_requirements.txt
coverage
pep8
pyflakes
pylint
pytest
pytest-cov
wheel
mock
flask-testing